Add event date as column to infectious history <data.frame>

The event date could be first contact, last contact or other.

Usage

.add_date_contact(
  .data,
  contact_type = c("first", "last"),
  distribution = c("pois", "geom"),
  ...,
  outbreak_start_date = NULL
)

.add_hospitalisation(.data, onset_to_hosp, hosp_risk)

.add_outcome(
  .data,
  onset_to_death,
  onset_to_recovery,
  hosp_death_risk,
  non_hosp_death_risk,
  config
)

Arguments

.data

A <data.frame> containing the infectious history from a branching process simulation.

contact_type

A character with the type of contact, either first contact ("first"), or last contact ("last").

distribution

A character with the name of the distribution, following the base R convention for distribution naming (e.g. Poisson is pois).

...

dots Extra arguments to be passed to the distribution function given in the distribution argument.

outbreak_start_date

A date for the start of the outbreak.

onset_to_hosp

An <epidist> object, an anonymous function for the onset to hospitalisation delay distribution, or NA to not simulate hospitalisation (admission) dates.

hosp_risk

Either a single numeric for the hospitalisation risk of everyone in the population, or a <data.frame> with age specific hospitalisation risks Default is 20% hospitalisation (0.2) for the entire population. If the onset_to_hosp argument is set to NA this argument should also be set to NA. See details and examples for more information.

onset_to_death

An <epidist> object, an anonymous function for the onset to death delay distribution, or NA to not simulate dates for individuals that died.

onset_to_recovery

An <epidist> object, an anonymous function for the onset to death delay distribution, or NA to not simulate dates for individuals that recovered. Default is NA so by default cases that recover get an NA in the $date_outcome line list column.

hosp_death_risk

Either a single numeric for the death risk for hospitalised individuals across the population, or a <data.frame> with age specific hospitalised death risks Default is 50% death risk in hospitals (0.5) for the entire population. If the onset_to_death argument is set to NA this argument should also be set to NA. See details and examples for more information. If a time-varying death risk is specified in the config the hosp_death_risk is interpreted as the maximum risk across the epidemic.

non_hosp_death_risk

Either a single numeric for the death risk for outside of hospitals across the population, or a <data.frame> with age specific death risks outside of hospitals. Default is 5% death risk outside of hospitals (0.05) for the entire population. If the onset_to_death argument is set to NA this argument should also be set to NA. See details and examples for more information. If a time-varying death risk is specified in the config the non_hosp_death_risk is interpreted as the maximum risk across the epidemic.

config

A list of settings to adjust the randomly sampled delays and Ct values. See create_config() for more information.

Value

A <data.frame> with one more column than input into .data. Unless the column heading is already present in which the data is overwritten.