Create a mock line list without simulating to test internal functions that require line list data

Create a mock line list without simulating to test internal functions that require line list data

Usage

.create_linelist(scenario)

Arguments

scenario

A character string with the state of the line list. Options are:

"pre_date_last_contact"

"pre_date_first_contact"

"pre_hospitalisation"

"pre_deaths"

"pre_names"

"pre_ct"

Value

A <data.frame>.

Details

Each of the scenarios are created by taking a snapshot (i.e copy) of the line list <data.frame> are each stage of the simulation (sim_linelist()) after setting the seed to 1 (set.seed(1)). The scenarios are before each function call, for example, "pre_date_last_contact" is the state of the line list prior to calling .add_date_last_contact(), and "pre_names" is the state of the line list prior to calling .add_names().

Script to reproduce data:

# load data required to simulate line list
serial_interval <- epiparameter::epidist(
  disease = "COVID-19",
  epi_dist = "serial interval",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)

# get onset to hospital admission from {epiparameter} database
onset_to_hosp <- epiparameter::epidist_db(
  disease = "COVID-19",
  epi_dist = "onset to hospitalisation",
  single_epidist = TRUE
)

# get onset to death from {epiparameter} database
onset_to_death <- epiparameter::epidist_db(
  disease = "COVID-19",
  epi_dist = "onset to death",
  single_epidist = TRUE
)
set.seed(1)

linelist <- sim_linelist(
  R = 1.1,
  serial_interval = serial_interval,
  onset_to_hosp = onset_to_hosp,
  onset_to_death = onset_to_death,
  hosp_rate = 0.5,
  add_ct = TRUE
)