Internal simulation function called by the exported simulation functions within simulist

This internal function simulates a line list, and when sim_type is "contacts" or "outbreak" a contacts table as well.

Usage

.sim_internal(
  sim_type = c("linelist", "contacts", "outbreak"),
  contact_distribution,
  infect_period,
  prob_infect,
  onset_to_hosp = NULL,
  onset_to_death = NULL,
  onset_to_recovery = NULL,
  hosp_risk = NULL,
  hosp_death_risk = NULL,
  non_hosp_death_risk = NULL,
  outbreak_start_date,
  anonymise = NULL,
  outbreak_size,
  population_age,
  case_type_probs = NULL,
  contact_tracing_status_probs = NULL,
  config
)

Arguments

sim_type

A character string specifying which simulation function this function is being called within.

contact_distribution

An <epidist> object or anonymous function for the contact distribution. This is any discrete density function that produces non-negative integers (including zero, \(\mathbb{N}_0\)) for the number of contacts per infection.

infect_period

An <epidist> object or anonymous function for the infectious period. This defines the duration from becoming infectious to no longer infectious. In the simulation, individuals are assumed to become infectious immediately after being infected (the latency period is assumed to be zero). The time intervals between an infected individual and their contacts are assumed to be uniformly distributed within the infectious period.

prob_infect

A single numeric for the probability of a secondary contact being infected by an infected primary contact.

onset_to_hosp

An <epidist> object, an anonymous function for the onset to hospitalisation delay distribution, or NA to not simulate hospitalisation (admission) dates.

onset_to_death

An <epidist> object, an anonymous function for the onset to death delay distribution, or NA to not simulate dates for individuals that died.

onset_to_recovery

An <epidist> object, an anonymous function for the onset to death delay distribution, or NA to not simulate dates for individuals that recovered. Default is NA so by default cases that recover get an NA in the $date_outcome line list column.

hosp_risk

Either a single numeric for the hospitalisation risk of everyone in the population, or a <data.frame> with age specific hospitalisation risks Default is 20% hospitalisation (0.2) for the entire population. If the onset_to_hosp argument is set to NA this argument should also be set to NA. See details and examples for more information.

hosp_death_risk

Either a single numeric for the death risk for hospitalised individuals across the population, or a <data.frame> with age specific hospitalised death risks Default is 50% death risk in hospitals (0.5) for the entire population. If the onset_to_death argument is set to NA this argument should also be set to NA. See details and examples for more information. If a time-varying death risk is specified in the config the hosp_death_risk is interpreted as the maximum risk across the epidemic.

non_hosp_death_risk

Either a single numeric for the death risk for outside of hospitals across the population, or a <data.frame> with age specific death risks outside of hospitals. Default is 5% death risk outside of hospitals (0.05) for the entire population. If the onset_to_death argument is set to NA this argument should also be set to NA. See details and examples for more information. If a time-varying death risk is specified in the config the non_hosp_death_risk is interpreted as the maximum risk across the epidemic.

outbreak_start_date

A date for the start of the outbreak.

anonymise

A logical boolean for whether case names should be anonymised. Default is FALSE.

outbreak_size

A numeric vector of length 2 defining the minimum and the maximum number of infected individuals for the simulated outbreak. Default is c(10, 1e4), so the minimum outbreak size is 10 infected individuals, and the maximum outbreak size is 10,000 infected individuals. Either number can be changed to increase or decrease the maximum or minimum outbreak size to allow simulating larger or smaller outbreaks. If the minimum outbreak size cannot be reached after running the simulation for many iterations (internally) then the function errors, whereas if the maximum outbreak size is exceeded the function returns the data early and a warning stating how many cases and contacts are returned.

population_age

Either a numeric vector with two elements or a <data.frame> with age structure in the population. Use a numeric vector to specific the age range of the population, the first element is the lower bound for the age range, and and the second is the upper bound for the age range (both inclusive, i.e. [lower, upper]). The <data.frame> with age groups and the proportion of the population in that group. See details and examples for more information.

case_type_probs

A named numeric vector with the probability of each case type. The names of the vector must be "suspected", "probable", "confirmed". Values of each case type must sum to one.

contact_tracing_status_probs

A named numeric vector with the probability of each contact tracing status. The names of the vector must be "under_followup", "lost_to_followup", "unknown". Values of each contact tracing status must sum to one.

config

A list of settings to adjust the randomly sampled delays and Ct values. See create_config() for more information.

Value

A <data.frame> if sim_type is "linelist" or "contacts", or a list of two <data.frame>s if sim_type is "outbreak".