Simulate a linelist

The linelist is simulated using a branching process and parameterised with previously published epidemiological parameters.

Usage

sim_linelist(
  R,
  serial_interval,
  onset_to_hosp,
  onset_to_death,
  hosp_rate = 0.2,
  hosp_death_rate = 0.5,
  non_hosp_death_rate = 0.05,
  outbreak_start_date = as.Date("2023-01-01"),
  add_names = TRUE,
  add_ct = FALSE,
  min_chain_size = 10,
  age_range = c(1, 90),
  case_type_probs = c(suspected = 0.2, probable = 0.3, confirmed = 0.5),
  include_contacts = FALSE,
  ...
)

Arguments

R

A single numeric for the reproduction number.

serial_interval

An <epidist> object for the serial interval.

onset_to_hosp

An <epidist> object for the onset to hospitalisation delay distribution.

onset_to_death

An <epidist> object for the onset to death delay distribution.

hosp_rate

Either a single numeric for the hospitalisation rate of everyone in the population, or a <data.frame> with age specific hospitalisation rates. Default is 20% hospitalisation (0.2) for the entire population. See details and examples for more information.

hosp_death_rate

Either a single numeric for the death rate for hospitalised individuals across the population, or a <data.frame> with age specific hospitalised death rates. Default is 50% death rate in hospitals (0.5) for the entire population. See details and examples for more information.

non_hosp_death_rate

Either a single numeric for the death rate for outside of hospitals across the population, or a <data.frame> with age specific death rates outside of hospitals. Default is 5% death rate outside of hospitals (0.05) for the entire population. See details and examples for more information.

outbreak_start_date

A date for the start of the outbreak.

add_names

A logical boolean for whether to add names to each row of the linelist. Default is TRUE.

add_ct

A logical boolean for whether to add Ct values to each row of the linelist. Default is FALSE.

min_chain_size

A single numeric defining the minimum chain size for the simulated outbreak. Default is 10. This can be increased when the function should simulate a larger outbreak.

age_range

A numeric vector with two elements. The first is the lower bound for the age range, and and the second is the upper bound for the age range (both inclusive, i.e. [lower, upper]).

case_type_probs

A named numeric vector with the probability of each case type. The names of the vector must be "suspected", "probable", "confirmed". Values of each case type must sum to one.

include_contacts

A logical boolean to determine whether contacts are added to the linelist. Currently not implmented.

...

dots Extra arguments to be passed to other functions. Currently not used and will return a warning if extra arguments are supplied.

Value

A linelist <data.frame>

Details

For age-stratified hospitalised and death rates a <data.frame> will need to be passed to the hosp_rate and/or hosp_death_rate arguments. This <data.frame> should have three columns:

min_age: a column with one numeric per cell for the minimum age of the age group (inclusive).

max_age: a column with one numeric per cell for the maximum age of the age group (exclusive).

rate: a column with one numeric per cell for the proportion (or probability) of hospitalisation for that age group. Should be between 0 and 1.

Author

Joshua W. Lambert, Carmen Tamayo

Examples

# load data required to simulate linelist
serial_interval <- epiparameter::epidist(
  disease = "COVID-19",
  epi_dist = "serial interval",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)
#> Citation cannot be created as author, year, journal or title is missing

# get onset to hospital admission from {epiparameter} database
onset_to_hosp <- epiparameter::epidist_db(
  disease = "COVID-19",
  epi_dist = "onset to hospitalisation",
  single_epidist = TRUE
)
#> Using Linton NM, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov AR, Jung S,
#> Yuan B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.. 
#> To retrieve the short citation use the 'get_citation' function

# get onset to death from {epiparameter} database
onset_to_death <- epiparameter::epidist_db(
  disease = "COVID-19",
  epi_dist = "onset to death",
  single_epidist = TRUE
)
#> Using Linton NM, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov AR, Jung S,
#> Yuan B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.. 
#> To retrieve the short citation use the 'get_citation' function
# example with single hospitalisation rate for entire population
linelist <- sim_linelist(
  R = 1.1,
  serial_interval = serial_interval,
  onset_to_hosp = onset_to_hosp,
  onset_to_death = onset_to_death,
  hosp_rate = 0.5
)

# example with age-stratified hospitalisation rate
# 20% for over 80s
# 10% for under 5s
# 5% for the rest
age_dep_hosp_rate <- data.frame(
  min_age = c(1, 5, 80),
  max_age = c(5, 80, 90),
  rate = c(0.1, 0.05, 0.2)
)
linelist <- sim_linelist(
  R = 1.1,
  serial_interval = serial_interval,
  onset_to_hosp = onset_to_hosp,
  onset_to_death = onset_to_death,
  hosp_rate = age_dep_hosp_rate
)