Internal simulation function called by the exported simulation functions within simulist

This internal function simulates a line list, and when sim_type is "contacts" or "outbreak" a contacts table as well.

Usage

.sim_internal(
  sim_type = c("linelist", "contacts", "outbreak"),
  contact_distribution,
  contact_interval,
  prob_infect,
  onset_to_hosp = NULL,
  onset_to_death = NULL,
  hosp_risk = NULL,
  hosp_death_risk = NULL,
  non_hosp_death_risk = NULL,
  outbreak_start_date,
  add_names = NULL,
  add_ct = NULL,
  min_outbreak_size,
  population_age,
  case_type_probs = NULL,
  contact_tracing_status_probs = NULL,
  config
)

Arguments

sim_type

A character string specifying which simulation function this function is being called within.

contact_distribution

An <epidist> object or anonymous function for the contact distribution. This is any discrete density function that produces non-negative integers (including zero, \(\mathbb{N}_0\)) for the number of contacts per infection.

contact_interval

An <epidist> object or anonymous function for the contact interval. This is analogous to the serial interval or generation time, but defines the time interval between an individual being infected/infectious (in the simulation the latency period is assumed to be zero) and having contact with another susceptible individual.

prob_infect

A single numeric for the probability of a secondary contact being infected by an infected primary contact.

onset_to_hosp

An <epidist> object or anonymous function for the onset to hospitalisation delay distribution.

onset_to_death

An <epidist> object or anonymous function for the onset to death delay distribution.

hosp_risk

Either a single numeric for the hospitalisation risk of everyone in the population, or a <data.frame> with age specific hospitalisation risks Default is 20% hospitalisation (0.2) for the entire population. See details and examples for more information.

hosp_death_risk

Either a single numeric for the death risk for hospitalised individuals across the population, or a <data.frame> with age specific hospitalised death risks Default is 50% death risk in hospitals (0.5) for the entire population. See details and examples for more information.

non_hosp_death_risk

Either a single numeric for the death risk for outside of hospitals across the population, or a <data.frame> with age specific death risks outside of hospitals. Default is 5% death risk outside of hospitals (0.05) for the entire population. See details and examples for more information.

outbreak_start_date

A date for the start of the outbreak.

add_names

A logical boolean for whether to add names to each row of the line list. Default is TRUE.

add_ct

A logical boolean for whether to add Ct values to each confirmed case and NA otherwise for each case in the line list. Default is TRUE. Ct refers to the Cycle threshold from a Real-time PCR or quantitative PCR (qPCR).

min_outbreak_size

A single numeric defining the minimum chain size for the simulated outbreak. Default is 10. This can be increased when the function should simulate a larger outbreak.

population_age

Either a numeric vector with two elements or a <data.frame> with age structure in the population. Use a numeric vector to specific the age range of the population, the first element is the lower bound for the age range, and and the second is the upper bound for the age range (both inclusive, i.e. [lower, upper]). The <data.frame> with age groups and the proportion of the population in that group. See details and examples for more information.

case_type_probs

A named numeric vector with the probability of each case type. The names of the vector must be "suspected", "probable", "confirmed". Values of each case type must sum to one.

contact_tracing_status_probs

A named numeric vector with the probability of each contact tracing status. The names of the vector must be "under_followup", "lost_to_followup", "unknown". Values of each contact tracing status must sum to one.

config

A list of settings to adjust the randomly sampled delays and Ct values (if add_ct = TRUE). See create_config() for more information.

Value

A <data.frame> if sim_type is "linelist" or "contacts", or a list of two <data.frame>s if sim_type is "outbreak".