Simulate a line list

The line list is simulated using a branching process and parameterised with user defined epidemiological parameters.

Usage

sim_linelist(
  R,
  serial_interval,
  onset_to_hosp,
  onset_to_death,
  hosp_risk = 0.2,
  hosp_death_risk = 0.5,
  non_hosp_death_risk = 0.05,
  outbreak_start_date = as.Date("2023-01-01"),
  add_names = TRUE,
  add_ct = TRUE,
  min_outbreak_size = 10,
  population_age = c(1, 90),
  case_type_probs = c(suspected = 0.2, probable = 0.3, confirmed = 0.5),
  config = create_config()
)

Arguments

R

A single numeric for the reproduction number.

serial_interval

An <epidist> object or anonymous function for the serial interval.

onset_to_hosp

An <epidist> object or anonymous function for the onset to hospitalisation delay distribution.

onset_to_death

An <epidist> object or anonymous function for the onset to death delay distribution.

hosp_risk

Either a single numeric for the hospitalisation risk of everyone in the population, or a <data.frame> with age specific hospitalisation risks Default is 20% hospitalisation (0.2) for the entire population. See details and examples for more information.

hosp_death_risk

Either a single numeric for the death risk for hospitalised individuals across the population, or a <data.frame> with age specific hospitalised death risks Default is 50% death risk in hospitals (0.5) for the entire population. See details and examples for more information.

non_hosp_death_risk

Either a single numeric for the death risk for outside of hospitals across the population, or a <data.frame> with age specific death risks outside of hospitals. Default is 5% death risk outside of hospitals (0.05) for the entire population. See details and examples for more information.

outbreak_start_date

A date for the start of the outbreak.

add_names

A logical boolean for whether to add names to each row of the line list. Default is TRUE.

add_ct

A logical boolean for whether to add Ct values to each row of the line list. Default is TRUE.

min_outbreak_size

A single numeric defining the minimum chain size for the simulated outbreak. Default is 10. This can be increased when the function should simulate a larger outbreak.

population_age

Either a numeric vector with two elements or a <data.frame> with age structure in the population. Use a numeric vector to specific the age range of the population, the first element is the lower bound for the age range, and and the second is the upper bound for the age range (both inclusive, i.e. [lower, upper]). The <data.frame> with age groups and the proportion of the population in that group. See details and examples for more information.

case_type_probs

A named numeric vector with the probability of each case type. The names of the vector must be "suspected", "probable", "confirmed". Values of each case type must sum to one.

config

A list of settings to adjust the randomly sampled delays and Ct values (if add_ct = TRUE). See create_config() for more information.

Value

A line list <data.frame>

Details

For age-stratified hospitalised and death risks a <data.frame> will need to be passed to the hosp_risk and/or hosp_death_risk arguments. This <data.frame> should have two columns:

age_limit: a column with one numeric per cell for the lower bound (minimum) age of the age group (inclusive).

risk: a column with one numeric per cell for the proportion (or probability) of hospitalisation for that age group. Should be between 0 and 1.

For an age structured population, a <data.frame> with two columns:

age_range: a column with characters specifying the lower and upper bound of that age group, separated by a hyphen (-). Both bounds are inclusive (integers). For example, an age group of one to ten would be given as "1-10".

proportion: a column with the proportion of the population that are in that age group. Proportions must sum to one.

Author

Joshua W. Lambert, Carmen Tamayo

Examples

# load data required to simulate line list
serial_interval <- epiparameter::epidist(
  disease = "COVID-19",
  epi_dist = "serial interval",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)
#> Citation cannot be created as author, year, journal or title is missing

# get onset to hospital admission from {epiparameter} database
onset_to_hosp <- epiparameter::epidist_db(
  disease = "COVID-19",
  epi_dist = "onset to hospitalisation",
  single_epidist = TRUE
)
#> Using Linton N, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov A, Jung S, Yuan
#> B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.. 
#> To retrieve the short citation use the 'get_citation' function

# get onset to death from {epiparameter} database
onset_to_death <- epiparameter::epidist_db(
  disease = "COVID-19",
  epi_dist = "onset to death",
  single_epidist = TRUE
)
#> Using Linton N, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov A, Jung S, Yuan
#> B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.. 
#> To retrieve the short citation use the 'get_citation' function
# example with single hospitalisation risk for entire population
linelist <- sim_linelist(
  R = 1.1,
  serial_interval = serial_interval,
  onset_to_hosp = onset_to_hosp,
  onset_to_death = onset_to_death,
  hosp_risk = 0.5
)
head(linelist)
#>   id        case_name case_type gender age date_onset date_admission date_death
#> 1  1 Ashraf el-Hafeez suspected      m   6 2023-01-01     2023-01-06       <NA>
#> 2  2  Miranda Caffery  probable      f  23 2023-01-01           <NA>       <NA>
#> 3  3      Edward Veed  probable      m  59 2023-01-01     2023-01-06       <NA>
#> 4  4  Sabaaha al-Mady  probable      f  33 2023-01-02     2023-01-05       <NA>
#> 5  5     Marquis York confirmed      m  83 2023-01-01     2023-01-06       <NA>
#> 6  6       Mariah Tom suspected      f  29 2023-01-02     2023-01-02       <NA>
#>   date_first_contact date_last_contact ct_value
#> 1               <NA>              <NA>       NA
#> 2         2023-01-04        2023-01-04       NA
#> 3         2023-01-04        2023-01-05       NA
#> 4         2022-12-29        2023-01-02       NA
#> 5         2023-01-01        2023-01-02     24.4
#> 6         2022-12-29        2023-01-03       NA

# example with age-stratified hospitalisation risk
# 20% for over 80s
# 10% for under 5s
# 5% for the rest
age_dep_hosp_risk <- data.frame(
  age_limit = c(1, 5, 80),
  risk = c(0.1, 0.05, 0.2)
)
linelist <- sim_linelist(
  R = 1.1,
  serial_interval = serial_interval,
  onset_to_hosp = onset_to_hosp,
  onset_to_death = onset_to_death,
  hosp_risk = age_dep_hosp_risk
)
head(linelist)
#>   id           case_name case_type gender age date_onset date_admission
#> 1  1        Melanie Tang confirmed      f  50 2023-01-01           <NA>
#> 2  2    Anthony Franklin suspected      m  21 2023-01-01           <NA>
#> 3  3    Alexis Mcalister confirmed      f  71 2023-01-01           <NA>
#> 4  4 Amanda Baudendistel confirmed      f  25 2023-01-01           <NA>
#> 5  5       Jerod Roberts  probable      m   1 2023-01-01           <NA>
#> 6  6      Christin North  probable      f  13 2023-01-03           <NA>
#>   date_death date_first_contact date_last_contact ct_value
#> 1       <NA>               <NA>              <NA>     24.7
#> 2       <NA>         2023-01-07        2023-01-07       NA
#> 3       <NA>         2022-12-30        2023-01-04     24.7
#> 4       <NA>         2022-12-31        2023-01-03     24.7
#> 5       <NA>         2023-01-01        2023-01-06       NA
#> 6       <NA>         2023-01-02        2023-01-04       NA