Simulate a line list

The line list is simulated using a branching process and parameterised with user defined epidemiological parameters.

Usage

sim_linelist(
  contact_distribution,
  infect_period,
  prob_infect,
  onset_to_hosp,
  onset_to_death,
  onset_to_recovery = NA,
  hosp_risk = 0.2,
  hosp_death_risk = 0.5,
  non_hosp_death_risk = 0.05,
  outbreak_start_date = as.Date("2023-01-01"),
  add_names = TRUE,
  outbreak_size = c(10, 10000),
  population_age = c(1, 90),
  case_type_probs = c(suspected = 0.2, probable = 0.3, confirmed = 0.5),
  config = create_config()
)

Arguments

contact_distribution

An <epidist> object or anonymous function for the contact distribution. This is any discrete density function that produces non-negative integers (including zero, \(\mathbb{N}_0\)) for the number of contacts per infection.

infect_period

An <epidist> object or anonymous function for the infectious period. This defines the duration from becoming infectious to no longer infectious. In the simulation, individuals are assumed to become infectious immediately after being infected (the latency period is assumed to be zero). The time intervals between an infected individual and their contacts are assumed to be uniformly distributed within the infectious period.

prob_infect

A single numeric for the probability of a secondary contact being infected by an infected primary contact.

onset_to_hosp

An <epidist> object, an anonymous function for the onset to hospitalisation delay distribution, or NA to not simulate hospitalisation (admission) dates.

onset_to_death

An <epidist> object, an anonymous function for the onset to death delay distribution, or NA to not simulate dates for individuals that died.

onset_to_recovery

An <epidist> object, an anonymous function for the onset to death delay distribution, or NA to not simulate dates for individuals that recovered. Default is NA so by default cases that recover get an NA in the $date_outcome line list column.

hosp_risk

Either a single numeric for the hospitalisation risk of everyone in the population, or a <data.frame> with age specific hospitalisation risks Default is 20% hospitalisation (0.2) for the entire population. If the onset_to_hosp argument is set to NA this argument should also be set to NA. See details and examples for more information.

hosp_death_risk

Either a single numeric for the death risk for hospitalised individuals across the population, or a <data.frame> with age specific hospitalised death risks Default is 50% death risk in hospitals (0.5) for the entire population. If the onset_to_death argument is set to NA this argument should also be set to NA. See details and examples for more information. If a time-varying death risk is specified in the config the hosp_death_risk is interpreted as the maximum risk across the epidemic.

non_hosp_death_risk

Either a single numeric for the death risk for outside of hospitals across the population, or a <data.frame> with age specific death risks outside of hospitals. Default is 5% death risk outside of hospitals (0.05) for the entire population. If the onset_to_death argument is set to NA this argument should also be set to NA. See details and examples for more information. If a time-varying death risk is specified in the config the non_hosp_death_risk is interpreted as the maximum risk across the epidemic.

outbreak_start_date

A date for the start of the outbreak.

add_names

A logical boolean for whether to add names to each row of the line list. Default is TRUE.

outbreak_size

A numeric vector of length 2 defining the minimum and the maximum number of infected individuals for the simulated outbreak. Default is c(10, 1e4), so the minimum outbreak size is 10 infected individuals, and the maximum outbreak size is 10,000 infected individuals. Either number can be changed to increase or decrease the maximum or minimum outbreak size to allow simulating larger or smaller outbreaks. If the minimum outbreak size cannot be reached after running the simulation for many iterations (internally) then the function errors, whereas if the maximum outbreak size is exceeded the function returns the data early and a warning stating how many cases and contacts are returned.

population_age

Either a numeric vector with two elements or a <data.frame> with age structure in the population. Use a numeric vector to specific the age range of the population, the first element is the lower bound for the age range, and and the second is the upper bound for the age range (both inclusive, i.e. [lower, upper]). The <data.frame> with age groups and the proportion of the population in that group. See details and examples for more information.

case_type_probs

A named numeric vector with the probability of each case type. The names of the vector must be "suspected", "probable", "confirmed". Values of each case type must sum to one.

config

A list of settings to adjust the randomly sampled delays and Ct values. See create_config() for more information.

Value

A line list <data.frame>

Details

For age-stratified hospitalised and death risks a <data.frame> will need to be passed to the hosp_risk and/or hosp_death_risk arguments. This <data.frame> should have two columns:

age_limit: a column with one numeric per cell for the lower bound (minimum) age of the age group (inclusive).

risk: a column with one numeric per cell for the proportion (or probability) of hospitalisation for that age group. Should be between 0 and 1.

For an age structured population, a <data.frame> with two columns:

age_range: a column with characters specifying the lower and upper bound of that age group, separated by a hyphen (-). Both bounds are inclusive (integers). For example, an age group of one to ten would be given as "1-10".

proportion: a column with the proportion of the population that are in that age group. Proportions must sum to one.

Author

Joshua W. Lambert, Carmen Tamayo

Examples

# load data required to simulate line list
contact_distribution <- epiparameter::epidist(
  disease = "COVID-19",
  epi_dist = "contact distribution",
  prob_distribution = "pois",
  prob_distribution_params = c(mean = 2)
)
#> Citation cannot be created as author, year, journal or title is missing

infect_period <- epiparameter::epidist(
  disease = "COVID-19",
  epi_dist = "infectious period",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)
#> Citation cannot be created as author, year, journal or title is missing

# get onset to hospital admission from {epiparameter} database
onset_to_hosp <- epiparameter::epidist_db(
  disease = "COVID-19",
  epi_dist = "onset to hospitalisation",
  single_epidist = TRUE
)
#> Using Linton N, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov A, Jung S, Yuan
#> B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.. 
#> To retrieve the citation use the 'get_citation' function

# get onset to death from {epiparameter} database
onset_to_death <- epiparameter::epidist_db(
  disease = "COVID-19",
  epi_dist = "onset to death",
  single_epidist = TRUE
)
#> Using Linton N, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov A, Jung S, Yuan
#> B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.. 
#> To retrieve the citation use the 'get_citation' function
# example with single hospitalisation risk for entire population
linelist <- sim_linelist(
  contact_distribution = contact_distribution,
  infect_period = infect_period,
  prob_infect = 0.5,
  onset_to_hosp = onset_to_hosp,
  onset_to_death = onset_to_death,
  hosp_risk = 0.5
)
head(linelist)
#>   id          case_name case_type sex age date_onset date_admission   outcome
#> 1  1      Steven Dagher  probable   m  37 2023-01-01     2023-01-12      died
#> 2  3    Chantel Buckley  probable   f  35 2023-01-01     2023-01-01      died
#> 3  5 Hamza el-Mohiuddin confirmed   m  30 2023-01-04           <NA> recovered
#> 4  6   Laaiqa el-Nasser confirmed   f   5 2023-01-04     2023-01-04      died
#> 5  8          Sage Rutz confirmed   f  76 2023-01-04     2023-01-04      died
#> 6  9        Kaia Basham  probable   f  21 2023-01-04     2023-01-07      died
#>   date_outcome date_first_contact date_last_contact ct_value
#> 1   2023-01-16               <NA>              <NA>       NA
#> 2   2023-01-09         2023-01-01        2023-01-04       NA
#> 3         <NA>         2023-01-01        2023-01-03     23.5
#> 4   2023-01-17         2022-12-30        2023-01-03     23.5
#> 5   2023-01-14         2023-01-07        2023-01-08     23.5
#> 6   2023-01-17         2023-01-04        2023-01-07       NA

# example with age-stratified hospitalisation risk
# 20% for over 80s
# 10% for under 5s
# 5% for the rest
age_dep_hosp_risk <- data.frame(
  age_limit = c(1, 5, 80),
  risk = c(0.1, 0.05, 0.2)
)
linelist <- sim_linelist(
  contact_distribution = contact_distribution,
  infect_period = infect_period,
  prob_infect = 0.5,
  onset_to_hosp = onset_to_hosp,
  onset_to_death = onset_to_death,
  hosp_risk = age_dep_hosp_risk
)
head(linelist)
#>   id        case_name case_type sex age date_onset date_admission   outcome
#> 1  1    Kaitlyn Brown confirmed   f  35 2023-01-01           <NA> recovered
#> 2  3  Lajetta Washpun confirmed   f  38 2023-01-01     2023-01-06      died
#> 3  4 Nancy Laviolette confirmed   f  10 2023-01-01           <NA> recovered
#> 4  5 Numair el-Hallal confirmed   m  39 2023-01-02     2023-01-04 recovered
#> 5  6     Hsing Verdos confirmed   f  83 2023-01-01           <NA> recovered
#> 6  7         Aaron Le  probable   m  34 2023-01-01           <NA> recovered
#>   date_outcome date_first_contact date_last_contact ct_value
#> 1         <NA>               <NA>              <NA>     25.2
#> 2   2023-01-17         2022-12-30        2023-01-02     25.2
#> 3         <NA>         2022-12-29        2023-01-03     25.2
#> 4         <NA>         2023-01-03        2023-01-03     25.2
#> 5         <NA>         2023-01-02        2023-01-04     25.2
#> 6         <NA>         2022-12-30        2023-01-04       NA